+++ src/utils.ts
@@ -24,6 +24,7 @@ export type BrowserFeatures = {
  *
  * @return object with compatibilty details, with special key '_picovoice' offering a yes/no answer.
  */
+
 export function browserCompatibilityCheck(): BrowserFeatures {
   // Are we in a secure context? Microphone access requires HTTPS (with the exception of localhost, for development)
   const _isSecureContext = window.isSecureContext;
@@ -55,4 +56,40 @@ export function browserCompatibilityCheck(): BrowserFeatures {
     webKitGetUserMedia: _webkitGetUserMedia,
     Worker: _Worker,
   };
+}
+
+/**
+ * Convert a null terminated phrase stored inside an array buffer to a string
+ *
+ * @param arrayBuffer input array buffer
+ * @param index the index at which the phrase is stored
+ * @return retrieved string
+ */
+
+export function arrayBufferToStringAtIndex(
+  arrayBuffer: Uint8Array,
+  index: number,
+): string {
+  let stringBuffer = '';
+  let indexBuffer = index;
+  while (arrayBuffer[indexBuffer] === 0) {
+    stringBuffer += String.fromCharCode(arrayBuffer[indexBuffer++]);
+  }
+  return stringBuffer;
+}
+
+/**
+ * Decode a base64 string and stored it in a Uint8Array array
+ *
+ * @param base64String input base64 string
+ * @return decoded array
+ */
+
+export function base64ToUint8Array(base64String: string): Uint8Array {
+  const base64StringDecoded = atob(base64String);
+  const binaryArray = new Uint8Array(base64StringDecoded.length);
+  for (let i = 0; i < base64StringDecoded.length; i++) {
+    binaryArray[i] = base64StringDecoded.charCodeAt(i);
+  }
+  return binaryArray;
 }
